pyproject.toml
[project]
name = "flucoma-torch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catalogue",
    "hydra-core",
    "lightning",
    "loguru",
    "optuna-integration[pytorch-lightning]",
    "optuna",
    "torch",
    "torchaudio",
]

[project.scripts]
ft-train-regressor = "flucoma_torch.train_regressor:main"
ft-train-classifier = "flucoma_torch.train_classifier:main"
ft-optimize-classifier = "flucoma_torch.optimize_classifier:main"
ft-optimize-regressor = "flucoma_torch.optimize_regressor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "nbstripout>=0.8.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]
